ANDN instruction lookup table

+++ jolt-core/src/zkvm/instruction/virtual_change_divisor_w.rs
@@ -28,11 +28,11 @@ impl InstructionFlags for VirtualChangeDivisorW {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualChangeDivisorW> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         // Always treat as 32-bit values for W instructions
         (
             self.register_state.rs1 as u32 as u64,
-            self.register_state.rs2 as i32 as i64,
+            self.register_state.rs2 as i32 as i128,
         )
     }
